Switch backend to i32

+++ server/migrations/2023-05-24-054054_users/up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE users (
-  user_id BIGINT NOT NULL PRIMARY KEY,
+  user_id INT NOT NULL PRIMARY KEY,
   first_name VARCHAR NOT NULL,
   last_name VARCHAR NOT NULL,
   email VARCHAR UNIQUE NOT NULL,
